stubs1b: fold the 27 casts the new bodies brought in - 360 -> 333
The merge could not land because 14 new bodies carried new reinterpret_casts on top
of main's freshly-ratcheted 335. Every one of them was a mis-modeled declaration, so
all four categories folded by fixing the type, not by parking the cast. The ledger is
now 333 total with ZERO never-examined - the OPEN worklist is empty.

CDDrawWorkerHost::WrapCoord takes LONG*, not i32* (-18)
    DrawObjectCounts @0x15a712 hands it `lea ecx,[esp+0x20]` / `lea eax,[esp+0x24]`,
    the .right/.bottom of the same 4-dword rect it then passes to DrawCount @0x164380,
    which passes it straight to DrawTextA. That makes the rect a genuine Win32 RECT,
    so its fields are LONG; C++ keeps long* and int* distinct at identical width, and
    an i32* declaration forced a reinterpret at all six WwdObjMgr corner sites. The
    other ten callers only take the address of a coordinate local - a type the image
    cannot name - so they follow the declaration and nothing casts. Byte-neutral.

CMapStringToOb::Lookup already hands back CObject*& (-5)
    DDrawWorkerRegistry punned five CDDrawWorker* locals into CObject*& to call it.
    But afxcoll.h declares `BOOL Lookup(LPCTSTR, CObject*&)` - the base pointer binds
    with no cast at all, and CDDrawWorker : CLoadable : CWapObj : CObject is plain
    single inheritance, so recovering the derived type is a real static_cast that
    emits nothing. The same TU's other five scan sites were ALREADY spelled that way
    (MapTeardown, RemoveKeysEqual, SumSizesEqual, AnyValueMatches, LoadNamespace);
    all ten now agree. All four DispatchKeyed* and InstallTree stay 100.00 EXACT.

CStatusBarItem::Setup's arg 9 is a registry key, not an int (-3)
    The old note claimed retail's mangling proved `HH` - but that mangled name is our
    own label, so the argument cited itself. The real evidence: the sibling slot-11
    CSBI_Image::SetupImage takes the same 9th dword, retail hands it a DATA-relocated
    string literal (BuildStatusBarTabs @0x10001a `push 0x614168` =
    "GAME_STATUSBAR_TABZ_STATZTAB"), and it feeds that dword straight to the same
    CMapStringToOb::Lookup that CSBI_WellGoo::Setup calls with arg 9. The three retail
    slot-2 call sites push a literal 0 - a null key - which is why the rect-only
    bodies ignore it. Retyped across the family; all three bodies held their exact
    previous score (WellGoo 83.52, RectOnly 90.38, StatusBarItem 90.00).

CPlay::DrawDebugStatsFull's rect is a RECT, not the sprintf scratch buffer (-1)
    This scored 99.10% by aliasing a char[0x40] sprintf buffer as a RECT, which forced
    MSVC to colour the rect onto that slot the way retail's did. That was a mis-model,
    not a discovery - no dev declared a RECT over a scratch array. The overlay is the
    stack allocator reusing a dead array, which no source spelling expresses: the same
    slot is a sprintf destination earlier in the body (`lea edx,[esp+0x30]; push
    0x612718; call sprintf`), and being address-exposed is why one of lr's stores
    survives while the other three forward into dr. Honest `RECT lr;` local, 94.72%,
    @early-stop updated with the mechanism so it is not re-punned.

Also: s_dbgUnknown -> s_dbgNoCaps (it is the arm where the caps say neither
VIDEOMEMORY nor SYSTEMMEMORY), taking `Unknown ids` 2 -> 0; and the jump-table span
pattern note gains the counter-case it was missing - growing a span is MEASURED, not
assumed, with the three tileswitchlogic spans that got WORSE (SetActionCode 66.6 ->
47.2, Process 58.9 -> 47.4, MorphByTool 92.1 -> 53.9).

Gates: `gruntz build --full` exit 0. reinterpret_casts 335 -> 333 (ratchet re-blessed
downward). MAX fuzzy 78.73 -> 80.16; every MAX in the regression report is held or
preserved. The 6 REGRESS / 16 LOST rows are main's own rehomes and dissolutions read
against the pre-merge baseline unit (WarpIsPow2 -> imagepolyclip, EndDialog ->
savegame, CWwdSlot9cA dissolved), plus ordinary /O2 ripple from the retyped headers.

+++ include/DDrawMgr/DDrawWorkerHost.h
@@ -114,7 +114,15 @@ public:
     // advances it by the byte count ReadPlaneObjects returns.
     i32 RebuildPlanes(const char* base, i32 count);
     i32 ReadPlaneObjects(const PlaneObjectRecord* src);
-    void WrapCoord(i32* px, i32* py); // 0x00a000 wrap+transform a world coord
+    // 0x00a000 wrap+transform a world coord. The out-params are LONG*, not i32*:
+    // DrawObjectCounts (0x15a712) hands this `lea ecx,[esp+0x20]` / `lea eax,[esp+0x24]`
+    // - the .right/.bottom of the very rect it then passes to DrawCount (0x164380),
+    // which passes it on to DrawTextA. That makes the rect a genuine Win32 RECT, so
+    // its fields are LONG; C++ keeps long* and int* distinct at identical width, so an
+    // i32* declaration would force a reinterpret at all six WwdObjMgr corner sites.
+    // The remaining callers take the address of a plain coordinate local, whose type
+    // the image cannot name - so they follow the declaration, and nothing casts.
+    void WrapCoord(LONG* px, LONG* py);
     // `out` is the {x,y} pair the body fills (out[0]/out[1]); its one caller hands it a
     // Coord, so the pair type goes on the declaration instead of at the call.
     void SnapToTileCenter(struct Coord* out, i32 x, i32 y); // 0x0311e0 snap (x,y) to tile centre
